Move tests of parameters for ssh_config creation into own test file
to speed up testing when parallelisation is used.

- rename file of old tests
